Do not download the Monaco editor parts from CDN
Bundle the editor locally
