+++ web/package.json
@@ -69,6 +69,7 @@
     "jest-environment-jsdom": "^30.2.0",
     "jest-transform-stub": "^2.0.0",
     "mini-css-extract-plugin": "^2.10.0",
+    "monaco-editor-webpack-plugin": "^7.1.1",
     "neostandard": "^0.13.0",
     "prettier": "^3.8.1",
     "react-refresh": "^0.18.0",
